Add files via upload
This pretrained model for forced alignment using the Montreal Forced Aligner (MFA) is trained on the Corpus of Regional African American Language (CORAAL), and a unique pronouncing dictionary that takes into account AAL phonology/phonetics. It is the counterpart of coraal1.zip, which is trained using CORAAL and a standard pronouncing dictionary.